+++ src/interface/term.ml
@@ -373,7 +373,9 @@ module type Tff = sig
   type ty
   type ty_var
   type ty_const
-  (** The representation of term types, type variables, and type constants. *)
+  type ty_def
+  (** The representation of term types, type variables, and type constants,
+      and lastly type definitions. *)
 
   type 'a tag
   (** The type of tags used to annotate arbitrary terms. *)
@@ -475,7 +477,7 @@ module type Tff = sig
   val define_adt :
     ty_const -> ty_var list ->
     (path * (ty * path option) list) list ->
-    (Cstr.t * (ty * Const.t option) list) list
+    ty_def * (Cstr.t * (ty * Const.t option) list) list
   (** [define_aft t vars cstrs] defines the type constant [t], parametrised over
       the type variables [ty_vars] as defining an algebraic datatypes with constructors
       [cstrs]. [cstrs] is a list where each elements of the form [(name, l)] defines
@@ -502,7 +504,7 @@ module type Tff = sig
   *)
 
   val define_record :
-    ty_const -> ty_var list -> (path * ty) list -> Field.t list
+    ty_const -> ty_var list -> (path * ty) list -> ty_def * Field.t list
   (** Define a (previously abstract) type to be a record type, with the given fields. *)
 
   exception Wrong_type of t * ty
@@ -796,8 +798,131 @@ module type Ae_Bitv = sig
 
   val extract : int -> int -> t -> t
   (** Bitvector extraction, using in that order,
-      the start and then end the position of the
+      the start and then end positions of the
       bitvector to extract. *)
+
+  val repeat : int -> t -> t
+  (** Repetition of a bitvector. *)
+
+  val zero_extend : int -> t -> t
+  (** Extend the given bitvector with the given number of 0s. *)
+
+  val sign_extend : int -> t -> t
+  (** Extend the given bitvector with its most significant bit
+      repeated the given number of times. *)
+
+  val rotate_right : int -> t -> t
+  (** [rotate_right i x] means rotate bits of x to the right i times. *)
+
+  val rotate_left : int -> t -> t
+  (** [rotate_left i x] means rotate bits of x to the left i times. *)
+
+  val not : t -> t
+  (** Bitwise negation. *)
+
+  val and_ : t -> t -> t
+  (** Bitwise conjunction. *)
+
+  val or_ : t -> t -> t
+  (** Bitwise disjunction. *)
+
+  val nand : t -> t -> t
+  (** [nand s t] abbreviates [not (and_ s t)]. *)
+
+  val nor : t -> t -> t
+  (** [nor s t] abbreviates [not (or_ s t)]. *)
+
+  val xor : t -> t -> t
+  (** [xor s t] abbreviates [or_ (and_ s (not t)) (and_ (not s) t)]. *)
+
+  val xnor : t -> t -> t
+  (** [xnor s t] abbreviates [or_ (and_ s t) (and_ (not s) (not t))]. *)
+
+  val comp : t -> t -> t
+  (** Bitwise comparison. [comp s t] equals [#b1] iff [s] and [t]
+      are bitwise equal. *)
+
+  val neg : t -> t
+  (** Arithmetic complement on bitvectors.
+      Supposing an input bitvector of size [m] representing
+      an integer [k], the resulting term should represent
+      the integer [2^m - k]. *)
+
+  val add : t -> t -> t
+  (** Arithmetic addition on bitvectors, modulo the size of
+      the bitvectors (overflows wrap around [2^m] where [m]
+      is the size of the two input bitvectors). *)
+
+  val sub : t -> t -> t
+  (** Arithmetic substraction on bitvectors, modulo the size
+      of the bitvectors (2's complement subtraction modulo).
+      [sub s t] should be equal to [add s (neg t)]. *)
+
+  val mul : t -> t -> t
+  (** Arithmetic multiplication on bitvectors, modulo the size
+      of the bitvectors (see {!add}). *)
+
+  val udiv : t -> t -> t
+  (** Arithmetic euclidian integer division on bitvectors. *)
+
+  val urem : t -> t -> t
+  (** Arithmetic euclidian integer remainder on bitvectors. *)
+
+  val sdiv : t -> t -> t
+  (** Arithmetic 2's complement signed division.
+      (see smtlib's specification for more information). *)
+
+  val srem : t -> t -> t
+  (** Arithmetic 2's complement signed remainder (sign follows dividend).
+      (see smtlib's specification for more information). *)
+
+  val smod : t -> t -> t
+  (** Arithmetic 2's complement signed remainder (sign follows divisor).
+      (see smtlib's specification for more information). *)
+
+  val shl : t -> t -> t
+  (** Logical shift left. [shl t k] return the result of
+      shifting [t] to the left [k] times. In other words,
+      this should return the bitvector representing
+      [t * 2^k] (since bitvectors represent integers using
+      the least significatn bit in cell 0). *)
+
+  val lshr : t -> t -> t
+  (** Logical shift right. [lshr t k] return the result of
+      shifting [t] to the right [k] times. In other words,
+      this should return the bitvector representing
+      [t / (2^k)]. *)
+
+  val ashr : t -> t -> t
+  (** Arithmetic shift right, like logical shift right except that the most
+      significant bits of the result always copy the most significant
+      bit of the first argument*)
+
+  val ult : t -> t -> t
+  (** Boolean arithmetic comparison (less than).
+      [ult s t] should return the [true] term iff [s < t]. *)
+
+  val ule : t -> t -> t
+  (** Boolean arithmetic comparison (less or equal than). *)
+
+  val ugt : t -> t -> t
+  (** Boolean arithmetic comparison (greater than). *)
+
+  val uge : t -> t -> t
+  (** Boolean arithmetic comparison (greater or equal than). *)
+
+  val slt : t -> t -> t
+  (** Boolean signed arithmetic comparison (less than).
+      (See smtlib's specification for more information) *)
+
+  val sle : t -> t -> t
+  (** Boolean signed arithmetic comparison (less or equal than). *)
+
+  val sgt : t -> t -> t
+  (** Boolean signed arithmetic comparison (greater than). *)
+
+  val sge : t -> t -> t
+  (** Boolean signed arithmetic comparison (greater or equal than). *)
 
 end
 
@@ -1259,7 +1384,7 @@ module type Smtlib_Bitv = sig
   (** Repetition of a bitvector. *)
 
   val zero_extend : int -> t -> t
-  (** Extend the given bitvector with the given numer of 0. *)
+  (** Extend the given bitvector with the given number of 0s. *)
 
   val sign_extend : int -> t -> t
   (** Extend the given bitvector with its most significant bit
@@ -1293,7 +1418,7 @@ module type Smtlib_Bitv = sig
   (** [xnor s t] abbreviates [or_ (and_ s t) (and_ (not s) (not t))]. *)
 
   val comp : t -> t -> t
-  (** Bitwise comparison. [comp s t] equald [#b1] iff [s] and [t]
+  (** Bitwise comparison. [comp s t] equals [#b1] iff [s] and [t]
       are bitwise equal. *)
 
   val neg : t -> t
@@ -1327,11 +1452,11 @@ module type Smtlib_Bitv = sig
       (see smtlib's specification for more information). *)
 
   val srem : t -> t -> t
-  (** Arithmetic 2's coplement signed remainder (sign follows dividend).
+  (** Arithmetic 2's complement signed remainder (sign follows dividend).
       (see smtlib's specification for more information). *)
 
   val smod : t -> t -> t
-  (** Arithmetic 2's coplement signed remainder (sign follows divisor).
+  (** Arithmetic 2's complement signed remainder (sign follows divisor).
       (see smtlib's specification for more information). *)
 
   val shl : t -> t -> t