+++ src/typecheck/float.ml
@@ -200,7 +200,10 @@ module Smtlib2 = struct
                 `Reserved (
                   "completing interpretation of fp.to_bv in models",
                   `Term_cst (meta_to_bv F.to_ubv'))
-              | `Script _ -> `Not_found
+              | `Script _ ->
+                (* the regular case is handled later, because fp.to_ubv is
+                   an indexed identifier. *)
+                `Not_found
             end
           | "fp.to_sbv" ->
             begin match version with
@@ -208,7 +211,10 @@ module Smtlib2 = struct
                 `Reserved (
                   "completing interpretation of fp.to_sbv in models",
                   `Term_cst (meta_to_bv F.to_sbv'))
-              | `Script _ -> `Not_found
+              | `Script _ ->
+                (* the regular case is handled later, because fp.to_sbv is
+                   an indexed identifier. *)
+                `Not_found
             end
           | _ -> `Not_found
         end